# swagger/data-exchange-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0",
    "title": "Data Exchange"
  },
  "host": "api.smartcity.kmitl.io",
  "basePath": "/api/v1",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User_req"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/users/login": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "login",
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": []
          },
          {
            "userToken": []
          }
        ]
      }
    },
    "/users/{userName}/token": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Generate access token",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "operation success",
            "schema": {
              "type": "string",
              "example": "token"
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get user's access token",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "access token",
            "schema": {
              "type": "string",
              "example": "token"
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
    "/users/{userName}/thumbnail": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "update user profile picture",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "thumbnail",
            "type": "file",
            "description": "Profile image in png format only."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "userToken": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "User_req": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    },
    "userToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "User token"
    },
    "ticket": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Ticket"
    }
  }
}